+++ Analyzer/src/Make2LInputTrees.cc
@@ -31,9 +31,9 @@ void Make2LInputTrees::Loop(NTupleReader& tr, double, int maxevents, bool)
         my_histos["EventCounter"]->Fill( eventCounter );
 
         const auto& runtype             = tr.getVar<std::string>("runtype");
-//        const auto& GoodLeptons         = tr.getVec<std::pair<std::string, TLorentzVector>>("GoodLeptons");
+//        const auto& GoodLeptons         = tr.getVec<std::pair<std::string, utility::LorentzVector>>("GoodLeptons");
         const auto& GoodLeptonsCharge   = tr.getVec<int>("GoodLeptonsCharge");
-//        const auto& Jets                = tr.getVec<TLorentzVector>("Jets");
+//        const auto& Jets                = tr.getVec<utility::LorentzVector>("Jets");
         const auto& NGoodLeptons        = tr.getVar<int>("NGoodLeptons");
 //        const auto& onZ                 = tr.getVar<bool>("onZ");
         const auto& JetID               = tr.getVar<bool>("JetID");
@@ -41,8 +41,8 @@ void Make2LInputTrees::Loop(NTupleReader& tr, double, int maxevents, bool)
 //        const auto& HT_trigger_pt30     = tr.getVar<double>("HT_trigger_pt30");
 //        const auto& NGoodJets_pt30      = tr.getVar<int>("NGoodJets_pt30");
 
-        const auto& TwoLep_Mbl1              = tr.getVar<double>("TwoLep_Mbl1");
-        const auto& TwoLep_Mbl2              = tr.getVar<double>("TwoLep_Mbl2");
+        const auto& TwoLep_Mbl1              = tr.getVar<float>("TwoLep_Mbl1");
+        const auto& TwoLep_Mbl2              = tr.getVar<float>("TwoLep_Mbl2");
 //        const auto& GoodBJets_pt30           = tr.getVec<bool>("GoodBJets_pt30");
